added things
Server code now runs.
Debug mode is a thing
Debug mode added to javascript file. (still need to add button to
interface.)
Untitled.py is a base for the GUI for the testing of the code without
lights, or a heads up with lights. The newer Javascript should have a
light viewer too…